Add a new library, libclang-cxx
The current libclang.so exports only the symbols required for the stable
C api. This is opposed to libLLVM.so, which exports all the symbols
from the LLVM libraries, including those from the C++ API. This patch
adds libclang-cxx.so that is similar to libLLVM.so and exports all
symbols from the clang libraries. The motivation for this library is to
be used by Clang tools that use Clang's C++ api. They no longer need to
link against the individual static libraries.

Pending upstream change https://reviews.llvm.org/D50359.

Bug: http://b/111579848
Test: build.py on Linux and Darwin and check libclang-cxx shared library

Change-Id: I2787fc96afbf2b35702896ea79157e3917394dc5